# backend/openapi.json
{
  "components": {
    "schemas": {
      "AbundanceProjectionRequest": {
        "properties": {
          "ranking": {
            "const": "mean_abundance",
            "default": "mean_abundance",
            "title": "Ranking",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/AnalysisScope"
          },
          "topN": {
            "default": 20,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Topn",
            "type": "integer"
          }
        },
        "title": "AbundanceProjectionRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AD-Meta API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analysis-runs/{run_key}": {
      "get": {
        "operationId": "analysis_run_api_analysis_runs__run_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis Run",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/analysis-runs/{run_key}/artifacts/{artifact_key}/features/query": {
      "post": {
        "operationId": "scoped_analysis_features_api_analysis_runs__run_key__artifacts__artifact_key__features_query_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopedFeatureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisFeaturePageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scoped Analysis Features",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/analysis-runs/{run_key}/artifacts/{artifact_key}/projection-audits/{projection_kind}": {
      "post": {
        "operationId": "projection_audit_api_analysis_runs__run_key__artifacts__artifact_key__projection_audits__projection_kind__post",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "projection_kind",
            "required": true,
            "schema": {
              "enum": [
                "abundance",
                "composition",
                "ko_contribution",
                "boxplot",
                "heatmap",
                "detection",
                "differential_ko",
                "taxonomy",
                "taxonomy_sankey",
                "pca",
                "pcoa"
              ],
              "title": "Projection Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectionAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectionAuditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Projection Audit",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/analysis-runs/{run_key}/artifacts/{artifact_key}/projection-audits/{projection_kind}/metadata": {
      "post": {
        "operationId": "projection_audit_metadata_api_analysis_runs__run_key__artifacts__artifact_key__projection_audits__projection_kind__metadata_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "projection_kind",
            "required": true,
            "schema": {
              "enum": [
                "abundance",
                "composition",
                "ko_contribution",
                "boxplot",
                "heatmap",
                "detection",
                "differential_ko",
                "taxonomy",
                "taxonomy_sankey",
                "pca",
                "pcoa"
              ],
              "title": "Projection Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectionAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectionAuditMetadataResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Projection Audit Metadata",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/analysis-runs/{run_key}/artifacts/{artifact_key}/projection-audits/{projection_kind}/options/{field}": {
      "post": {
        "operationId": "projection_audit_options_api_analysis_runs__run_key__artifacts__artifact_key__projection_audits__projection_kind__options__field__post",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "projection_kind",
            "required": true,
            "schema": {
              "enum": [
                "abundance",
                "composition",
                "ko_contribution",
                "boxplot",
                "heatmap",
                "detection",
                "differential_ko",
                "taxonomy",
                "taxonomy_sankey",
                "pca",
                "pcoa"
              ],
              "title": "Projection Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "field",
            "required": true,
            "schema": {
              "title": "Field",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectionAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectionAuditOptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Projection Audit Options",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/analysis-runs/{run_key}/artifacts/{artifact_key}/projection-audits/{projection_kind}/rows": {
      "post": {
        "operationId": "projection_audit_rows_api_analysis_runs__run_key__artifacts__artifact_key__projection_audits__projection_kind__rows_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "projection_kind",
            "required": true,
            "schema": {
              "enum": [
                "abundance",
                "composition",
                "ko_contribution",
                "boxplot",
                "heatmap",
                "detection",
                "differential_ko",
                "taxonomy",
                "taxonomy_sankey",
                "pca",
                "pcoa"
              ],
              "title": "Projection Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectionAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectionAuditRowsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Projection Audit Rows",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/analysis-runs/{run_key}/artifacts/{artifact_key}/projections/abundance": {
      "post": {
        "operationId": "abundance_projection_api_analysis_runs__run_key__artifacts__artifact_key__projections_abundance_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbundanceProjectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbundanceProjectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abundance Projection",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/analysis-runs/{run_key}/artifacts/{artifact_key}/projections/{projection_kind}": {
      "post": {
        "operationId": "chart_projection_api_analysis_runs__run_key__artifacts__artifact_key__projections__projection_kind__post",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "projection_kind",
            "required": true,
            "schema": {
              "enum": [
                "composition",
                "ko_contribution",
                "boxplot",
                "heatmap",
                "detection",
                "differential_ko",
                "taxonomy",
                "taxonomy_sankey",
                "pca",
                "pcoa"
              ],
              "title": "Projection Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartProjectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartProjectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chart Projection",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/analysis-runs/{run_key}/artifacts/{artifact_key}/results": {
      "get": {
        "operationId": "complete_results_api_analysis_runs__run_key__artifacts__artifact_key__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sampleCode",
            "required": false,
            "schema": {
              "default": "",
              "title": "Samplecode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "phenotype",
            "required": false,
            "schema": {
              "default": "",
              "title": "Phenotype",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "featureId",
            "required": false,
            "schema": {
              "default": "",
              "title": "Featureid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "default": "sampleCode",
              "title": "Sortby",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortDirection",
            "required": false,
            "schema": {
              "default": "asc",
              "title": "Sortdirection",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteResultPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Results",
        "tags": [
          "analysis-runs"
        ]
      }
    },
...
    "/api/analysis-runs/{run_key}/samples/{sample_code}": {
      "get": {
        "operationId": "analysis_sample_api_analysis_runs__run_key__samples__sample_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_key",
            "required": true,
            "schema": {
              "title": "Run Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sample_code",
            "required": true,
            "schema": {
              "title": "Sample Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSampleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis Sample",
        "tags": [
          "analysis-runs"
        ]
      }
    },
...
  }
}